requirements.txt
streamlit>=1.45.0
pandas>=2.2.0
openai>=1.68.0
python-dotenv>=1.1.0
